feat: Team 도메인 연관관게 추가, Question 바뀐 부분 없음

+++ src/main/java/com/flower/question/domain/Question.java
@@ -4,10 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 @Entity
 @Table(name = "question")
@@ -17,7 +14,7 @@ public class Question {
 
     @Id
     @Column(name = "question_id")
-    private int id;
+    private Long id;
 
     @Column(unique = true)
     private String questionList;
